Make BibTeX code block narrower
Trying to make it fit without horizontal scrolling on GitHub pages.
This is a hopeless task because we can't predict how wide the user
will make the window, but making our BibTeX block narrower ought to
be better for a larger percentage of cases.  (Or not, and I'm just
wasting my time again.)